Move Navigraph Auth to shared systems
