+++ fbw-common/src/systems/instruments/src/EFB/Navigation/Navigation.tsx
@@ -38,7 +38,7 @@ import { NavigraphPage } from './Pages/NavigraphPage/NavigraphPage';
 import { getPdfUrl, LocalFilesPage } from './Pages/LocalFilesPage/LocalFilesPage';
 import { PinnedChartUI } from './Pages/PinnedChartsPage';
 import { useNavigraphAuth } from '../../react/navigraph';
-import { navigraphCharts } from '../../navigraph';
+import { navigraphCharts } from '@flybywiresim/fbw-sdk';
 
 export const navigationTabs: (PageLink & { associatedTab: NavigationTab })[] = [
   { name: 'Navigraph', alias: '', component: <NavigraphPage />, associatedTab: NavigationTab.NAVIGRAPH },
