+++ fbw-common/src/systems/instruments/src/EFB/Navigation/Pages/NavigraphPage/NavigraphChartUI.tsx
@@ -13,7 +13,7 @@ import { SelectGroup, SelectItem } from '../../../UtilComponents/Form/Select';
 import { SimpleInput } from '../../../UtilComponents/Form/SimpleInput/SimpleInput';
 import { ScrollableContainer } from '../../../UtilComponents/ScrollableContainer';
 import { ChartViewer } from '../../Navigation';
-import { navigraphCharts } from '../../../../navigraph';
+import { navigraphCharts } from '@flybywiresim/fbw-sdk';
 import { ChartCategory } from '@flybywiresim/fbw-sdk';
 
 export const NavigraphChartUI = () => {
