+++ fbw-common/src/systems/shared/src/navigraph/charts.ts
@@ -1,0 +1,3 @@
+import { getChartsAPI } from 'navigraph/charts';
+
+export const navigraphCharts = getChartsAPI();
